Computed net worth
